+++ pkg/cmd/task/task.go
@@ -6,6 +6,7 @@ import (
 	"github.com/lucassabreu/clockify-cli/pkg/cmd/task/done"
 	"github.com/lucassabreu/clockify-cli/pkg/cmd/task/edit"
 	"github.com/lucassabreu/clockify-cli/pkg/cmd/task/list"
+	quickadd "github.com/lucassabreu/clockify-cli/pkg/cmd/task/quick-add"
 	"github.com/lucassabreu/clockify-cli/pkg/cmdutil"
 	"github.com/spf13/cobra"
 )
@@ -20,6 +21,7 @@ func NewCmdTask(f cmdutil.Factory) *cobra.Command {
 
 	cmd.AddCommand(list.NewCmdList(f, nil))
 	cmd.AddCommand(add.NewCmdAdd(f, nil))
+	cmd.AddCommand(quickadd.NewCmdQuickAdd(f, nil))
 	cmd.AddCommand(edit.NewCmdEdit(f, nil))
 	cmd.AddCommand(del.NewCmdDelete(f, nil))
 	cmd.AddCommand(done.NewCmdDone(f, nil))